[UPDATE] data type collection Payment record

+++ src/main/java/com/banquito/corecobros/commission/dto/CollectionPaymentRecordDTO.java
@@ -8,8 +8,8 @@ import java.time.LocalDate;
 @Value
 @Builder
 public class CollectionPaymentRecordDTO {
-    private Long idCollection;
-    private Long itemCommissionId;
+    private Integer idCollection;
+    private Integer itemCommissionId;
     private String uniqueId;
     private BigDecimal ammount;
     private String type;
